requirements.txt
autopep8==2.0.1
click==8.1.3
Flask==2.2.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
mysql-connector-python==8.0.32
protobuf==3.20.3
pycodestyle==2.10.0
tomli==2.0.1
Werkzeug==2.2.2
